refactor: tweaks on the project arch for folders and files
